backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.12.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.13
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
stripe==10.12.0
httpx==0.27.0
resend==2.4.0
